Feature/re architect book read

+++ api/internal/domain/entity/books_progress.go
@@ -5,9 +5,9 @@ import (
 )
 
 type BooksProgress struct {
-	ID        uint `gorm:"primary_key"`
-	BooksId   uint `json:"books_id"`
-	Page      int  `json:"page"`
+	ID        uint64 `gorm:"primary_key"`
+	BooksId   uint64 `json:"books_id"`
+	Page      uint64 `json:"page"`
 	CreatedAt time.Time
 	UpdatedAt time.Time
 }
